Basic Template Setup And Process Log
Setting up a basic template for the game in Unity, adding test buttons to start layout out the design of the game.

Started a txt file for process logging to appear in my repository, which I hope will be informative of my process on each day/time that I work on the test.